+++ src/main/java/com/b1n_ry/yigd/config/TrySoftConfig.java
@@ -3,5 +3,6 @@ package com.b1n_ry.yigd.config;
 public enum TrySoftConfig {
     RADIUS,
     Y_UP,
-    CLOSEST_Y
+    CLOSEST_Y,
+    DISABLED
 }