+++ src/main/java/com/b1n_ry/yigd/config/XpStorageConfig.java
@@ -1,0 +1,7 @@
+package com.b1n_ry.yigd.config;
+
+public enum XpStorageConfig {
+    IN_GRAVE,
+    KEEP_XP,
+    OUTSIDE_GRAVE
+}